Consolidated test plate setup in DiagnosticRTqPCRDebug
