+++ diagnostic_rt_qpcr/libraries/diagnosticrtqpcrdebug/source.rb
@@ -1,15 +1,17 @@
 # typed: false
 # frozen_string_literal: true
+
 needs 'Collection Management/CollectionTransfer'
-needs 'Diagnostic RT-qPCR/DiagnosticRTqPCRHelper'
+# needs 'Diagnostic RT-qPCR/DiagnosticRTqPCRHelper'
 needs 'Diagnostic RT-qPCR/DataAssociationKeys'
 
 module DiagnosticRTqPCRDebug
   include CollectionTransfer
-  include DiagnosticRTqPCRHelper
+  # include DiagnosticRTqPCRHelper
   include DataAssociationKeys
 
   VERBOSE = false
+  PLATE = 'PCR Plate'
 
   def debug_parameters
     {
@@ -17,144 +19,90 @@ module DiagnosticRTqPCRDebug
       program_name: 'CDC_TaqPath_CG',
       debug_primer: [5, 8, 9], # rp, n1, n2,
       debug_template: 'template',
-      debug_items: [257,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256,
-                    256]
+      debug_item_ids: [257] + Array.new(22, 256)
     }
   end
 
-  def setup_test(operations)
+  # Populate all input plates with qPCR Reactions
+  #
+  # @param operations [OperationList]
+  # @param method [String]
+  # @return [void]
+  def setup_test_plates(operations:, method: nil)
     operations.each do |op|
-      setup_test_plate(collection: op.input('PCR Plate').collection)
-      op.set_input('Template', generate_samples(debug_parameters[:debug_items]))
+      setup_test_plate(collection: op.input(PLATE).collection, method: method)
+      op.set_input('Template', Item.find(debug_parameters[:debug_item_ids]))
     end
   end
 
-  # Populate test plate with qPCR Reactions and one no template control (NTC)
+  # Populate a collection with qPCR Reactions
   #
-  def setup_test_plate(collection:)
+  # @param collection [Collection]
+  # @param method [String]
+  # @return [void]
+  def setup_test_plate(collection:, method:)
     verbose = false
-    key = TEMPLATE_KEY
     collection.associate(:program_name, debug_parameters[:program_name])
     collection.associate(:group_size, debug_parameters[:group_size])
     qpcr_reaction = Sample.find_by_name('Test qPCR Reaction')
-    ntc_item = Item.find(258)
+
+    layout_generator = PlateLayoutGeneratorFactory.build(
+      group_size: 24,
+      method: :cdc_primer_layout
+    )
+
+    loop do
+      layout_group = layout_generator.next_group
+      break unless layout_group.present?
+
+      layout_group.each do |r, c|
+        collection.set(r, c, qpcr_reaction)
+        part = collection.part(r, c)
+        part.associate(MASTER_MIX_KEY, 'added')
+      end
+    end
+
+    if method == :master_mix
+      show_result(collection: collection) if verbose
+      inspect_data_associations(collection: collection) if verbose
+      return
+    end
+
     layout_generator = PlateLayoutGeneratorFactory.build(
       group_size: debug_parameters[:group_size],
       method: :cdc_sample_layout
     )
-    i = 0
-    loop do
-      layout_group = layout_generator.next_group
-      break unless layout_group.present?
-      layout_group.each do |r, c|
-        collection.set(r, c, qpcr_reaction)
-        next if i.positive?
-        part = collection.part(r, c)
-        inspect part, "part at #{[r, c]}" if verbose
-        part.associate(key, ntc_item)
-        inspect part.associations, "#{key} at #{[r, c]}" if verbose
-      end
-      i += 1
+    layout_group = layout_generator.next_group
+    layout_group.each do |r, c|
+      part = collection.part(r, c)
+      part.associate(TEMPLATE_KEY, 'added')
     end
+
     show_result(collection: collection) if verbose
-    inspect collection.parts.to_s if verbose
+    inspect_data_associations(collection: collection) if verbose
   end
 
-  def generate_samples(debug_item_ids)
-    debug_items = []
-    debug_item_ids.each do |id|
-      debug_items.push(Item.find(id))
-    end
-    debug_items
-  end
-
+  # Show all the non-empty wells of the test plate
+  # @todo figure out what you really want to show here
+  #
+  # @param collection [Collection]
+  # @return [void]
   def show_result(collection:)
     show do
+      title 'Test Plate Setup'
       table highlight_non_empty(collection)
     end
   end
 
+  # Inspect a subset of the parts and their data associations
+  #
+  # @param collection [Collection]
+  # @return [void]
+  def inspect_data_associations(collection:)
+    [[0, 0], [0, 3], [0, 8]].each do |r, c|
+      part = collection.part(r, c)
+      inspect part, "part at #{[r, c]}"
+      inspect part.associations, "data at #{[r, c]}"
+    end
+  end
 end
-
-
-
-#   def setup_test(operations)
-#     operations.each do |op|
-#       op.input('PCR Plate').set(collection: generate_debug_container)
-#       op.set_input('Template', generate_samples(debug_parameters[:debug_items]))
-#     end
-#   end
-
-#   # TODO: this is pretty gross and hard coded yuck
-#   def generate_debug_container
-#     obj_type = ObjectType.find_by_name('96-well qPCR Plate')
-#     collection = Collection.new_collection(obj_type)
-#     collection.associate(:program_name, debug_parameters[:program_name])
-#     collection.associate(:group_size, debug_parameters[:group_size])
-
-#     add_primer_probe(collection)
-#     add_neg_control(collection)
-
-#     raise "the associations are #{collection.part(0,0).associations}"
-#     collection
-#   end
-
-#   def add_primer_probe(collection)
-#     debug_parameters[:debug_primer].each_with_index do |sample, idx|
-#       sample = Sample.find(sample)
-#       12.times do |col|
-#         row1 = 0 + idx
-#         row2 = 4 + idx
-#         collection.set(row1, col, sample)
-#         collection.set(row2, col, sample)
-#       end
-#     end
-#   end
-
-#   def add_neg_control(collection)
-#     microtiter_plate = MicrotiterPlateFactory.build(
-#       collection: collection,
-#       group_size: debug_parameters[:group_size],
-#       method: :sample_layout
-#     )
-#     item = Item.find(debug_parameters[:neg_control_sample])
-#     neg_group = microtiter_plate.next_empty(key: TEMPLATE_KEY)
-
-#     association_map = []
-#     neg_group.each { |r, c| association_map.push({ to_loc: [r, c] }) }
-#     transfer_from_item_to_collection(from_item: item,
-#                                      to_collection: microtiter_plate.collection,
-#                                      association_map: association_map,
-#                                      transfer_vol: { qty: 5, units: 'ul'})
-#     neg_group.each { |nxt| associate(index: nxt, key: TEMPLATE_KEY, data: item.id) }
-#   end
-
-#   def generate_samples(debug_item_ids)
-#     debug_items = []
-#     debug_item_ids.each do |id|
-#       debug_items.push(Item.find(id))
-#     end
-#     debug_items
-#   end
-# end
